chore: basic merge of functionality from menus.js
